Modify setting text size method

+++ kcsimplealertview/src/main/java/wooooooak/com/kcsimplealertview/model/AlertViewButtonModel.kt
@@ -1,10 +1,8 @@
 package wooooooak.dev.kcsimplealertview.woakalertview.model
 
-import androidx.annotation.DimenRes
-
 class AlertViewButtonModel(
     var text: String = "",
-    override var textSize: Int = 17,
+    override var textSize: Float = 17f,
     override var paddingTop: Int = 30,
     override var paddingBottom: Int = 30,
     var backgroundColor: Int = 0,
